+++ Tactical/XML_Magazine.cpp
@@ -2,44 +2,8 @@
 	#include "Tactical All.h"
 #else
 	#include "sgp.h"
-	#include "overhead types.h"
-	#include "Sound Control.h"
-	#include "Soldier Control.h"
 	#include "overhead.h"
-	#include "Event Pump.h"
 	#include "weapons.h"
-	#include "Animation Control.h"
-	#include "sys globals.h"
-	#include "Handle UI.h"
-	#include "Isometric Utils.h"
-	#include "worldman.h"
-	#include "math.h"
-	#include "points.h"
-	#include "ai.h"
-	#include "los.h"
-	#include "renderworld.h"
-	#include "opplist.h"
-	#include "interface.h"
-	#include "message.h"
-	#include "campaign.h"
-	#include "items.h"
-	#include "weapons.h"
-	#include "text.h"
-	#include "Soldier Profile.h"
-	#include "tile animation.h"
-	#include "Dialogue Control.h"
-	#include "SkillCheck.h"
-	#include "explosion control.h"
-	#include "Quests.h"
-	#include "Physics.h"
-	#include "Random.h"
-	#include "Vehicles.h"
-	#include "bullets.h"
-	#include "morale.h"
-	#include "meanwhile.h"
-	#include "SkillCheck.h"
-	#include "gamesettings.h"
-	#include "SaveLoadMap.h"
 	#include "Debug Control.h"
 	#include "expat.h"
 	#include "XML.h"
@@ -53,13 +17,13 @@ struct
 	MAGTYPE		curMagazine;
 	MAGTYPE *	curArray;
 	UINT32			maxArraySize;
-	
+
 	UINT32			currentDepth;
 	UINT32			maxReadDepth;
 }
 typedef magazineParseData;
 
-static void XMLCALL 
+static void XMLCALL
 magazineStartElementHandle(void *userData, const XML_Char *name, const XML_Char **atts)
 {
 	magazineParseData * pData = (magazineParseData *)userData;
@@ -105,11 +69,11 @@ magazineCharacterDataHandle(void *userData, const XML_Char *str, int len)
 {
 	magazineParseData * pData = (magazineParseData *)userData;
 
-	if( (pData->currentDepth <= pData->maxReadDepth) && 
+	if( (pData->currentDepth <= pData->maxReadDepth) &&
 		(strlen(pData->szCharData) < MAX_CHAR_DATA_LENGTH)
-	  ){
+	){
 		strncat(pData->szCharData,str,__min((unsigned int)len,MAX_CHAR_DATA_LENGTH-strlen(pData->szCharData)));
-	  }
+	}
 }
 
 
@@ -136,22 +100,22 @@ magazineEndElementHandle(void *userData, const XML_Char *name)
 		else if(strcmp(name, "uiIndex") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMagazine.uiIndex   = (UINT32) atol(pData->szCharData);
+			pData->curMagazine.uiIndex	= (UINT32) atol(pData->szCharData);
 		}
 		else if(strcmp(name, "ubCalibre") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMagazine.ubCalibre  = (UINT8) atol(pData->szCharData);
+			pData->curMagazine.ubCalibre	= (UINT8) atol(pData->szCharData);
 		}
 		else if(strcmp(name, "ubMagSize") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMagazine.ubMagSize = (UINT8) atol(pData->szCharData);
+			pData->curMagazine.ubMagSize = (UINT16) atol(pData->szCharData);
 		}
 		else if(strcmp(name, "ubAmmoType") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMagazine.ubAmmoType   = (UINT8) atol(pData->szCharData);
+			pData->curMagazine.ubAmmoType	= (UINT8) atol(pData->szCharData);
 		}
 
 		pData->maxReadDepth--;
@@ -170,7 +134,7 @@ BOOLEAN ReadInMagazineStats(STR fileName)
 	UINT32		uiFSize;
 	CHAR8 *		lpcBuffer;
 	XML_Parser	parser = XML_ParserCreate(NULL);
-	
+
 	magazineParseData pData;
 
 	DebugMsg(TOPIC_JA2, DBG_LEVEL_3, "Loading Magazines.xml" );
@@ -179,7 +143,7 @@ BOOLEAN ReadInMagazineStats(STR fileName)
 	hFile = FileOpen( fileName, FILE_ACCESS_READ, FALSE );
 	if ( !hFile )
 		return( FALSE );
-	
+
 	uiFSize = FileGetSize(hFile);
 	lpcBuffer = (CHAR8 *) MemAlloc(uiFSize+1);
 
@@ -194,19 +158,19 @@ BOOLEAN ReadInMagazineStats(STR fileName)
 
 	FileClose( hFile );
 
-	
+
 	XML_SetElementHandler(parser, magazineStartElementHandle, magazineEndElementHandle);
 	XML_SetCharacterDataHandler(parser, magazineCharacterDataHandle);
 
-	
+
 	memset(&pData,0,sizeof(pData));
 	pData.curArray = Magazine;
-	pData.maxArraySize = MAXITEMS; 
-	
+	pData.maxArraySize = MAXITEMS;
+
 	XML_SetUserData(parser, &pData);
 
 
-    if(!XML_Parse(parser, lpcBuffer, uiFSize, TRUE))
+	if(!XML_Parse(parser, lpcBuffer, uiFSize, TRUE))
 	{
 		CHAR8 errorBuf[511];
 
@@ -235,7 +199,7 @@ BOOLEAN WriteMagazineStats()
 	hFile = FileOpen( "TABLEDATA\\Magazine out.xml", FILE_ACCESS_WRITE | FILE_CREATE_ALWAYS, FALSE );
 	if ( !hFile )
 		return( FALSE );
-	
+
 	{
 		UINT32 cnt;
 
@@ -246,9 +210,9 @@ BOOLEAN WriteMagazineStats()
 			FilePrintf(hFile,"\t<MAGAZINE>\r\n");
 
 			FilePrintf(hFile,"\t\t<uiIndex>%d</uiIndex>\r\n",								cnt );
-			FilePrintf(hFile,"\t\t<ubCalibre>%d</ubCalibre>\r\n",								Magazine[cnt].ubCalibre  );
-			FilePrintf(hFile,"\t\t<ubMagSize>%d</ubMagSize>\r\n",								Magazine[cnt].ubMagSize   );
-			FilePrintf(hFile,"\t\t<ubAmmoType>%d</ubAmmoType>\r\n",								Magazine[cnt].ubAmmoType   );
+			FilePrintf(hFile,"\t\t<ubCalibre>%d</ubCalibre>\r\n",								Magazine[cnt].ubCalibre	);
+			FilePrintf(hFile,"\t\t<ubMagSize>%d</ubMagSize>\r\n",								Magazine[cnt].ubMagSize	);
+			FilePrintf(hFile,"\t\t<ubAmmoType>%d</ubAmmoType>\r\n",								Magazine[cnt].ubAmmoType	);
 
 			FilePrintf(hFile,"\t</MAGAZINE>\r\n");
 		}
